requirements.txt
requests==2.26.0
google-auth==2.3.1